move function suite into functions directory
